+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - Added a reporting tool to the control panel.
+- Added the **ModifyZipUrlFilesEvent** event class.
+- Added the **LinkClickEvent** event class.
 
 ### Fixed
 - Fixed numerous deprecation errors in control panel templates.
@@ -11,9 +13,12 @@
 - Fixed the installation migration to properly drop the custom fields table on uninstall.
 - Fixed a bug where element arrays were treated as objects.
 
+### Changed
+- Changed the name of the **FrontEndController** to **LinkVaultController**.
+
 ### Removed
-- Removed the LinkVaultDownload element index page.
-- Removed searchable columns from LinkVaultDownload element.
+- Removed the **LinkVaultDownload** element index page.
+- Removed searchable columns from **LinkVaultDownload** element.
 
 ## 3.0.5.7 - 2020-04-07